+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - feat: train_test_split (_split.py). Fixed Project Structure
 - feat: LinearRegrssion (_LinearRegression.py).  added README.md in linear_model
 - feat: LogisticRegression (_LogisticRegression.py). updated README.md in linear_model
+- feat: r2_score (_regression.py). added README.md in metrics